vibeseek/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pdf2genz"
version = "0.1.0"
description = "Local-first pipeline for converting PDF into short Gen Z videos."
requires-python = ">=3.11"
dependencies = [
  "pdfplumber>=0.11.0",
  "pydantic>=2.8.0",
  "python-dotenv>=1.0.1",
  "requests>=2.32.0",
  "tenacity>=8.5.0"
]

[tool.setuptools.packages.find]
where = ["src"]

[project.scripts]
pdf2genz = "pdf2genz.cli:main"

